Deprecated eust, eurostat added and TOC refactored
